+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `clear`: Clears the terminal screen.
 - `File::read_to_bytes`: Reads the whole file into a byte vector.
 
+### Changed
+
+- Implement the standard input, standard output, and standard error streams.
+
 ## [0.1.0-beta.2] - 2025-05-27
 
 ### Added